# api/cmd/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DatabaseInfo": {
        "description": "A managed PostgreSQL database (a postgres-16 family Firecracker sandbox). Connection fields are populated only while the database is running.",
        "properties": {
          "always_on": {
            "description": "true = opted out of idle auto-suspend.",
            "type": "boolean"
          },
          "broker_token": {
            "description": "Per-database REST query-broker bearer token (pds_pg_…); only returned while running.",
            "type": "string"
          },
          "broker_url": {
            "description": "Public REST query-broker base URL (/v1/databases/{id}/proxy).",
            "type": "string"
          },
          "cloned_from": {
            "description": "Source database id for clones/PITR restores.",
            "type": "string"
          },
          "branched_from": {
            "description": "Parent database id for reflink branches (TUSK T4.1).",
            "type": "string"
          },
          "connection_url": {
            "description": "Ready-to-use postgres:// URL (TLS required).",
            "type": "string"
          },
          "pooled_connection_url": {
            "description": "postgres:// URL routed through the guest's pgbouncer (transaction pooling) on port 6432. Recommended for serverless/agent workloads that open many short connections.",
            "type": "string"
          },
          "created_at": {
            "description": "Creation time, epoch seconds. Omitted when unknown.",
            "format": "int64",
            "type": "integer"
          },
          "database": {
            "description": "Default database name.",
            "type": "string"
          },
          "error": {
            "description": "Set when the VM is up but Postgres has not published credentials, or as a human-readable progress hint on async 202 responses.",
            "type": "string"
          },
          "failover_available": {
            "description": "Only on failed databases: whether a restore on another agent is possible.",
            "type": "boolean"
          },
          "failover_eta_seconds": {
            "description": "Expected restore time (RTO) in seconds when failover is available.",
            "type": "integer"
          },
          "failover_reason": {
            "description": "Why failover is or is not available.",
            "type": "string"
          },
          "host": {
            "description": "SNI host for native connections (<id>.db.pandastack.ai).",
            "type": "string"
          },
          "id": {
            "description": "Database id (currently doubles as the sandbox id).",
            "type": "string"
          },
          "label": {
            "description": "User-supplied display label.",
            "type": "string"
          },
          "password": {
            "description": "PostgreSQL password (rotates on restore/failover and on reset-credentials).",
            "type": "string"
          },
          "port": {
            "description": "PostgreSQL port (5432).",
            "type": "integer"
          },
          "sandbox_id": {
            "description": "Underlying sandbox id (explicit DB↔sandbox correlation).",
            "type": "string"
          },
          "size": {
            "description": "RAM tier, derived from the template.",
            "enum": [
              "1g",
              "4g",
              "16g"
            ],
            "type": "string"
          },
          "status": {
            "description": "Public status: provisioning | running | hibernated | failed | restoring (internal \"creating\" maps to provisioning).",
            "type": "string"
          },
          "template": {
            "description": "Backing template: postgres-16.",
            "type": "string"
          },
          "username": {
            "description": "PostgreSQL user.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "template"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "description": "Human-readable error message",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "pds_... API token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "url": "https://pandastack.ai"
    },
    "description": "REST API for PandaStack — Firecracker microVM sandboxes, managed Postgres, git-driven apps, serverless functions, and outbound webhooks. Authenticate with a pds_ bearer token from the dashboard's API Tokens page.",
    "title": "PandaStack API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/databases": {
      "get": {
        "operationId": "listDatabases",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "count": {
                      "description": "Number of items.",
                      "type": "integer"
                    },
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/DatabaseInfo"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "items",
                    "count"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Databases in the workspace (connection fields are not included in list items)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Workspace not set."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Agent unreachable."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List all managed databases in the caller's workspace",
        "tags": [
          "Databases"
        ]
      },
      "post": {
        "operationId": "createDatabase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseCreateRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseInfo"
                }
              }
            },
            "description": "Database sandbox accepted. Poll GET /v1/databases/{id} until status is \"running\" (connection credentials appear then)."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid size (valid: 1g, 4g, 16g)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Workspace not set (unauthenticated)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Failed to decode the agent's sandbox response."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Agent error while creating the underlying sandbox."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Provision a new managed PostgreSQL 16 database (async; returns 202 with status \"provisioning\")",
        "tags": [
          "Databases"
        ]
      }
    },
    "/v1/databases/{id}": {
      "delete": {
        "operationId": "deleteDatabase",
        "parameters": [
          {
            "description": "Database id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Database deleted; backup archive cleanup runs best-effort."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Database not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A clone of this database is still provisioning from its backups — retry once the clone is running."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Agent delete failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Database host transiently unreachable (agent restarting) — retry shortly. Includes Retry-After: 15.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Destroy a database and purge its GCS backup archive (irreversible)",
        "tags": [
          "Databases"
        ]
      },
      "get": {
        "operationId": "getDatabase",
        "parameters": [
          {
            "description": "Database id (doubles as the sandbox id).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseInfo"
                }
              }
            },
            "description": "Database details. When the VM is up but Postgres has not published credentials yet, status reports \"provisioning\" with an error hint. A failed database includes failover_available/failover_reason/failover_eta_seconds."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Database not found (or the sandbox is not a managed-database template)."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Agent unreachable."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a database with connection info (credentials populated once running)",
        "tags": [
          "Databases"
        ]
      },
      "patch": {
        "operationId": "updateDatabase",
        "parameters": [
          {
            "description": "Database id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "always_on": {
                    "description": "true opts the database out of idle auto-suspend.",
                    "type": "boolean"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseInfo"
                }
              }
            },
            "description": "Refreshed database view reflecting the update."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid body, or no updatable fields provided (supported: always_on)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Database not found."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Config update failed on the agent."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update mutable database config (currently only always_on, the idle auto-suspend opt-out)",
        "tags": [
          "Databases"
        ]
      }
    },
...
    "/v1/databases/{id}/failover": {
      "post": {
        "operationId": "failoverDatabase",
        "parameters": [
          {
            "description": "Database id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "force": {
                    "description": "true turns the healthy-primary refusal into a planned migration. Default false.",
                    "type": "boolean"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseInfo"
                }
              }
            },
            "description": "Failover started in the background (status \"restoring\"). Poll GET /v1/databases/{id}; the database may briefly report not-found while the restore provisions. Password rotates on restore."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Workspace not set."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Database not found (or not owned by this workspace)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Without force: database is running (use force for planned migration), hibernated (use wake), or in another non-failed state; or a failover for this database is already in progress."
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No restorable GCS archive — primary left untouched."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sandbox lookup failed."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Failover requires a multi-node deployment."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No healthy agent available to fail over to (primary left untouched)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Rebuild a failed database on a healthy agent from its GCS archive (preflight-then-act, async 202)",
        "tags": [
          "Databases"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://api.pandastack.ai"
    }
  ]
}